add statements, fix some bugs.
I edited the checkNext* functions so that the "advance" won't be inside them, because it created a lot of problems. Because of that, I tried to edit every usage of it- but I might have been mistaken in some of your functions, because I know them less. so now it works fine on ExpressionLessSquare's Main till it reaches the "if" statement. If you can check it that'll be great. Please pay attention to all the changes I've made (you can enter the commit itself and you'll see the additions to the file in green, deletations in red.